fix(sensors): CP3 review — log visibility, error-code truth, mock UB, test hardening
Review findings A1-A5, B1-B9, C1-C4+C8: infrastructure I2C failures now
ESP_LOGE/WARN instead of compiled-out debug (wedged bus distinguishable
from absent device); bounded init-failure logging (sensorless boot no
longer WARNs every 5 s); NaN pre-first-read placeholders (self-announcing
for non-gating consumers); MockEnvironmentalSensor initialize() reports
error codes like the real driver; MockI2cBus wrap-around UB fixed;
EspI2cBus handle table mutex-guarded for PR-05 multi-task sharing; env
command concurrent-state hint; error-code docs corrected (mid-init error
2, loss sequence 2-then-1); 10 new host tests (init-failure branches,
chip-ID fallthrough, digP1=0 guard, nonzero H3 / negative H5-H6 vectors,
clamps, Locked delegation, foreign chip-ID) — suite 119/0, rev1+rev2
green.

+++ firmware/components/sensors/include/sensors/testing/MockI2cBus.h
@@ -69,13 +69,14 @@ public:
     }
 
     /// Script consecutive register bytes starting at @p startReg (the device
-    /// is added implicitly if absent; the block must fit below 0x100).
+    /// is added implicitly if absent; indices wrap at 0xFF, matching
+    /// readRegisters()).
     void setRegisters(uint8_t address7, uint8_t startReg,
                       const std::vector<uint8_t>& values)
     {
         auto& map = devices_[address7];
         for (size_t i = 0; i < values.size(); ++i) {
-            map[startReg + i] = values[i];
+            map[static_cast<uint8_t>(startReg + i)] = values[i];
         }
     }
